TEST prise en compte des commentaires de la PR
