+++ mapie/tests/test_classification.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from copy import deepcopy
-from typing import Any, Dict, Iterable, List, Optional, Union, cast
+from typing import Any, Dict, Iterable, Optional, Union, cast
 
 import numpy as np
 import pandas as pd
@@ -25,10 +25,6 @@ from mapie._typing import ArrayLike, NDArray
 from mapie.classification import MapieClassifier
 from mapie.metrics import classification_coverage_score
 from mapie.utils import check_alpha
-
-from mapie.conformity_scores.utils_classification_conformity_scores import (
-    get_true_label_position
-)
 
 random_state = 42
 
@@ -59,25 +55,6 @@ Y_PRED_PROBA_WRONG = [
             [0.9, 0.01, 0.03, 0.06]
         ]
     )
-]
-
-Y_TRUE_PROBA_PLACE = [
-    [
-        np.array([2, 0]),
-        np.array([
-            [.1, .3, .6],
-            [.2, .7, .1]
-        ]),
-        np.array([[0], [1]])
-    ],
-    [
-        np.array([1, 0]),
-        np.array([
-            [.7, .12, .18],
-            [.5, .24, .26]
-        ]),
-        np.array([[2], [0]])
-    ]
 ]
 
 Params = TypedDict(
@@ -1856,22 +1833,6 @@ def test_get_last_included_proba_shape(k_lambda, strategy):
     assert y_p_p_c.shape == (len(X), len(np.unique(y)), len(thresholds))
     assert y_p_i_l.shape == (len(X), 1, len(thresholds))
     assert y_p_p_i_l.shape == (len(X), 1, len(thresholds))
-
-
-@pytest.mark.parametrize("y_true_proba_place", Y_TRUE_PROBA_PLACE)
-def test_get_true_label_position(
-    y_true_proba_place: List[NDArray]
-) -> None:
-    """
-    Check that the returned true label position the good.
-    """
-    y_true = y_true_proba_place[0]
-    y_pred_proba = y_true_proba_place[1]
-    place = y_true_proba_place[2]
-
-    found_place = get_true_label_position(y_pred_proba, y_true)
-
-    assert (found_place == place).all()
 
 
 @pytest.mark.parametrize("cv", [5, None])
